+++ client/app/components/Backlog/_styled.js
@@ -2,6 +2,7 @@ import styled from 'styled-components';
 
 import {
   COLOR_BACKGROUND_GREY,
+  COLOR_BLUE,
   COLOR_LIGHT_GREY,
   COLOR_LIGHTER_GREY,
   COLOR_ORANGE,
@@ -57,7 +58,12 @@ export const StyledEditForm = styled.form`
   }
 
   textarea {
-    min-height: 64px;
+    min-height: 128px;
+    margin-top: 2px;
+
+    &:focus {
+      border-bottom: 1px solid ${COLOR_BLUE};
+    }
   }
 `;
 
@@ -72,6 +78,10 @@ export const StyledAddForm = styled(StyledEditForm)`
   margin: 0 8px;
   padding: 8px 16px 16px 16px;
   border: 1px solid ${COLOR_LIGHTER_GREY};
+
+  textarea {
+    min-height: 64px;
+  }
 `;
 
 export const StyledBacklogInfoText = styled.div`
@@ -131,8 +141,12 @@ export const StyledStory = styled.div`
   box-sizing: border-box;
   border: 1px solid ${COLOR_LIGHTER_GREY};
   cursor: pointer;
-  border-left: ${({selected}) =>
-    selected ? '2px solid ' + COLOR_ORANGE : '1px solid ' + COLOR_LIGHTER_GREY};
+  border-left: ${({selected, highlighted}) =>
+    selected
+      ? '2px solid ' + COLOR_ORANGE
+      : highlighted
+      ? '1px solid ' + COLOR_BLUE
+      : '1px solid ' + COLOR_LIGHTER_GREY};
 
   &:hover {
     box-shadow: ${({noShadow}) =>
@@ -177,11 +191,6 @@ export const StyledStoryToolbar = styled.div`
   }
 `;
 
-export const StyledStoryText = styled.div`
-  overflow-x: hidden;
-  white-space: pre-wrap;
-`;
-
 export const StyledHighlightButtonWrapper = styled.div`
   width: 100%;
   display: flex;
@@ -197,7 +206,7 @@ export const StyledStoryAttributes = styled.div`
   }
 `;
 
-export const StyledBacklogSortForm = styled.form`
+export const StyledBacklogToolbar = styled.form`
   padding: 8px 8px 8px 8px;
   margin-right: 8px;
   border-bottom: 1px solid ${COLOR_LIGHTER_GREY};
@@ -216,11 +225,26 @@ export const StyledBacklogSortForm = styled.form`
   }
 
   > i {
+    display: block;
+    min-width: 16px;
+    text-align: center;
+  }
+
+  > i.icon-exchange {
     margin-left: 4px;
+    margin-right: 8px;
     transform: rotate(90deg);
+  }
+
+  > i.icon-trash:hover {
+    color: ${COLOR_WARNING};
   }
 `;
 
 export const StyledSortDropdownItem = styled.div`
   color: ${({selected}) => (selected ? COLOR_LIGHTER_GREY : 'inherit')};
+
+  > i {
+    margin-right: 4px;
+  }
 `;