+++ client/app/state/ui/uiReducer.js
@@ -1,9 +1,16 @@
 import getDayOfYear from 'date-fns/getDayOfYear';
 
 import {EVENT_ACTION_TYPES} from '../actions/eventActions';
-import {getHideNewUserHints, setHideNewUserHints} from '../clientSettingsStore';
+import {
+  getHideNewUserHints,
+  getMarkdownEnabled,
+  setHideNewUserHints,
+  setMarkdownEnabled
+} from '../clientSettingsStore';
 import {
   BACKLOG_SIDEBAR_TOGGLED,
+  MARKDOWN_TOGGLED,
+  MATRIX_TOGGLED,
   NEW_USER_HINTS_HIDDEN,
   SIDEBAR_ACTIONLOG,
   SIDEBAR_TOGGLED
@@ -13,10 +20,12 @@ import {LOCATION_CHANGED} from '../actions/commandActions';
 export const uiInitialState = {
   backlogShown: false, // only relevant in mobile view. in desktop the backlog is always visible and not "toggleable"
   sidebar: undefined,
+  matrixShown: false,
   applause: false,
   unseenError: false,
   easterEggActive: isHalloweenSeason(),
-  newUserHintHidden: getHideNewUserHints()
+  newUserHintHidden: getHideNewUserHints(),
+  markdownEnabled: getMarkdownEnabled()
 };
 
 /**
@@ -42,8 +51,14 @@ export default function uiReducer(state = uiInitialState, action, ownUserId) {
         return state;
       }
     }
-    case EVENT_ACTION_TYPES.storySelected:
-      return {...state, applause: false};
+    case EVENT_ACTION_TYPES.storySelected: {
+      if (action.event.userId === ownUserId) {
+        // if story was selected by ourself, make sure to hide matrix / show defaul story view
+        return {...state, applause: false, matrixShown: false};
+      } else {
+        return {...state, applause: false};
+      }
+    }
     case EVENT_ACTION_TYPES.consensusAchieved:
       return {...state, applause: true};
     case EVENT_ACTION_TYPES.newEstimationRoundStarted:
@@ -67,6 +82,13 @@ export default function uiReducer(state = uiInitialState, action, ownUserId) {
       }
     }
 
+    case MATRIX_TOGGLED: {
+      return {
+        ...state,
+        matrixShown: !state.matrixShown
+      };
+    }
+
     case BACKLOG_SIDEBAR_TOGGLED: {
       const showBacklog = !state.backlogShown;
 
@@ -80,6 +102,10 @@ export default function uiReducer(state = uiInitialState, action, ownUserId) {
     case NEW_USER_HINTS_HIDDEN: {
       setHideNewUserHints(true);
       return {...state, newUserHintHidden: true};
+    }
+    case MARKDOWN_TOGGLED: {
+      setMarkdownEnabled(!state.markdownEnabled);
+      return {...state, markdownEnabled: !state.markdownEnabled};
     }
     case LOCATION_CHANGED: {
       return {...state, pathname: action.pathname};