+++ client/cypress/integration/multiUser.spec.js
@@ -1,4 +1,4 @@
-import {v4 as uuid} from 'uuid';
+import {nanoid} from 'nanoid';
 import {tid} from '../support/commands';
 import {Landing, Room} from '../elements/elements';
 
@@ -9,7 +9,7 @@ beforeEach(function () {
 });
 
 it('multi user estimation with two rounds, consensus on second round', function () {
-  const roomId = 'multi-user-e2e_' + uuid();
+  const roomId = 'multi-user-e2e_' + nanoid().toLowerCase();
 
   cy.visit('/' + roomId);
 
@@ -21,8 +21,8 @@ it('multi user estimation with two rounds, consensus on second round', function 
   // -- the following will open an additional socket to the backend, so that we can add another user to the room
   // Cypress does not support multiple browsers!   see    https://docs.cypress.io/guides/references/trade-offs.html
   // see ../support/commands.js
-  const userTwoSocket = uuid();
-  const userTwoUserId = uuid();
+  const userTwoSocket = nanoid();
+  const userTwoUserId = nanoid().toLowerCase();
   cy.openNewSocket(userTwoSocket);
   cy.sendCommands(userTwoSocket, [
     {
@@ -103,7 +103,7 @@ it('multi user estimation with two rounds, consensus on second round', function 
 });
 
 it('estimation summary and settling on a value', function () {
-  const roomId = 'multi-user-e2e_' + uuid();
+  const roomId = 'multi-user-e2e_' + nanoid().toLowerCase();
 
   cy.visit('/' + roomId);
 
@@ -112,8 +112,8 @@ it('estimation summary and settling on a value', function () {
   Room.TopBar.whoami().click();
   Room.TopBar.whoamiDropdown().contains(this.user.username);
 
-  const userTwoSocket = uuid();
-  const userTwoUserId = uuid();
+  const userTwoSocket = nanoid();
+  const userTwoUserId = nanoid().toLowerCase();
   cy.openNewSocket(userTwoSocket);
   cy.sendCommands(userTwoSocket, [
     {
@@ -159,7 +159,7 @@ it('estimation summary and settling on a value', function () {
       // user one (e.g. the scrum master) settles on "3"
       Room.EstimationArea.summaryCard(3).click();
 
-      cy.get(tid('estimationSummary')).contains('Manually settled on "3"');
+      cy.get(tid('estimationSummary')).contains('Manually settled on');
       Room.EstimationArea.storyConsensus().contains('3');
       Room.Users.userEstimationGivenRevealed(3);
       Room.Users.userEstimationGivenRevealed(5);