+++ server/test/system/poinzSocketClient.js
@@ -1,5 +1,6 @@
 import socketIoClient from 'socket.io-client';
-import {v4 as uuid} from 'uuid';
+
+import uuid from '../../src/uuid';
 
 /**
  * provides functionality to interact with our poinz backend over websockets.
@@ -44,9 +45,8 @@ export default function poinzSocketClientFactory(backendUrl = 'http://localhost:
       clearEstimate,
       newRound,
       setCardConfig,
-      toggleConfidence,
-      toggleAutoReveal,
-      setPassword
+      setPassword,
+      setRoomConfig
     }
   };
 
@@ -319,20 +319,12 @@ export default function poinzSocketClientFactory(backendUrl = 'http://localhost:
     });
   }
 
-  function toggleAutoReveal(roomId, userId) {
+  function setRoomConfig(roomId, userId, config) {
     return sendCommand({
-      name: 'toggleAutoReveal',
+      name: 'setRoomConfig',
       roomId,
       userId,
-      payload: {}
-    });
-  }
-  function toggleConfidence(roomId, userId) {
-    return sendCommand({
-      name: 'toggleConfidence',
-      roomId,
-      userId,
-      payload: {}
+      payload: config
     });
   }